Unify geom packages.
