+++ modules/core/jvm/src/main/scala/lucuma/core/math/geom/jts/jvm/syntax/Awt.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.core.math.geom.jts
+package lucuma.core.geom.jts
 package jvm.syntax
 
 import org.locationtech.jts.awt.ShapeWriter
