+++ modules/core/shared/src/main/scala/lucuma/core/geom/GmosOiwfsProbeArm.scala
@@ -5,9 +5,9 @@ package lucuma.core.geom
 
 import lucuma.core.math.{ Angle, Offset }
 import lucuma.core.math.syntax.int._
-import lucuma.core.math.geom.syntax.all._
+import lucuma.core.geom.syntax.all._
 import lucuma.core.enum.{ GmosNorthFpu, GmosSouthFpu, PortDisposition }
-import lucuma.core.math.geom.ShapeExpression
+import lucuma.core.geom.ShapeExpression
 import cats.implicits._
 import scala.math.{ Pi, asin, atan2, hypot, sin }
