+++ modules/core/shared/src/main/scala/lucuma/core/geom/ShapeInterpreter.scala
@@ -1,10 +1,11 @@
 // Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.core.geom.jts.jvm
+package lucuma.core.geom
 
-package object syntax {
-
-  object all extends ToAwtOps
-
+/**
+  * Interprets a `ShapeExpression` to a `Shape`.
+  */
+trait ShapeInterpreter {
+  def interpret(e: ShapeExpression): Shape
 }
