+++ modules/core/shared/src/main/scala/lucuma/core/geom/syntax/package.scala
@@ -1,10 +1,12 @@
 // Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.core.geom.jts.jvm
+package lucuma.core.geom
 
 package object syntax {
 
-  object all extends ToAwtOps
+  // Syntax used in the JTS / JVM implementation only.
+
+  object all extends ToShapeExpressionOps with ToShapeExpressionCompanionOps
 
 }
