--- modules/core/shared/src/main/scala/lucuma/core/math/geom/ShapeInterpreter.scala
@@ -1,12 +1,0 @@
-// Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package lucuma.core.math.geom
-
-/**
- * Interprets a `ShapeExpression` to a `Shape`.
- */
-trait ShapeInterpreter {
-  def interpret(e: ShapeExpression): Shape
-}
-
