--- modules/core/shared/src/main/scala/lucuma/core/math/geom/jts/syntax/package.scala
@@ -1,14 +1,0 @@
-// Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package lucuma.core.math.geom.jts
-
-package object syntax {
-
-  // Syntax used in the JTS implementation only.
-
-  object all extends ToAngleOps
-                with ToOffsetOps
-
-}
-
